+++ src/main/java/com/nnk/springboot/services/RatingService.java
@@ -9,6 +9,9 @@ import org.springframework.stereotype.Service;
 import com.nnk.springboot.domain.Rating;
 import com.nnk.springboot.repositories.RatingRepository;
 
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.Valid;
+
 @Service
 public class RatingService {
 
@@ -23,11 +26,28 @@ public class RatingService {
 		return ratingRepository.findById(id);
 	}
 
-	public Rating saveRating(Rating rating) {
+	public Rating saveRating(@Valid Rating rating) {
 		return ratingRepository.save(rating);
+	}
+	
+	
+	public Rating updateRating(Integer id, @Valid Rating updatedRating) {
+		return ratingRepository.findById(id)
+				.map(rating -> {
+					rating.setMoodys_rating(updatedRating.getMoodys_rating());
+					rating.setSandprating(updatedRating.getSandprating());
+					rating.setFitch_rating(updatedRating.getFitch_rating());
+					rating.setOrder_number(updatedRating.getOrder_number());
+					return ratingRepository.save(rating);
+				})
+				.orElseThrow(() -> new EntityNotFoundException("Rating with id " + id + " not found"));
 	}
 
 	public void deleteRating(Integer id) {
-		ratingRepository.deleteById(id);
+		if(ratingRepository.existsById(id)) {
+			ratingRepository.deleteById(id);
+		} else {
+			throw new EntityNotFoundException("Rating with id " + id + " not found");
+		}
 	}
 }
