+++ src/main/java/com/nnk/springboot/services/TradeService.java
@@ -9,6 +9,9 @@ import org.springframework.stereotype.Service;
 import com.nnk.springboot.domain.Trade;
 import com.nnk.springboot.repositories.TradeRepository;
 
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.Valid;
+
 @Service
 public class TradeService {
 
@@ -23,11 +26,26 @@ public class TradeService {
 		return tradeRepository.findById(trade_id);
 	}
 
-	public Trade saveTrade(Trade trade) {
+	public Trade saveTrade(@Valid Trade trade) {
 		return tradeRepository.save(trade);
+	}
+	
+	public Trade updateTrade(Integer trade_id, @Valid Trade updatedTrade) {
+		return tradeRepository.findById(trade_id)
+				.map(trade -> {
+					trade.setAccount(updatedTrade.getAccount());
+					trade.setType(updatedTrade.getType());
+					trade.setBuy_quantity(updatedTrade.getBuy_quantity());
+					return tradeRepository.save(trade);
+				})
+				.orElseThrow(() -> new EntityNotFoundException("Trade with id " + trade_id + " not found"));
 	}
 
 	public void deleteTrade(Integer trade_id) {
-		tradeRepository.deleteById(trade_id);
+		if(tradeRepository.existsById(trade_id)) {
+			tradeRepository.deleteById(trade_id);
+		} else {
+			throw new EntityNotFoundException("Trade with id " + trade_id + " not found");
+		}
 	}
 }
